requirements.txt
deep-translator
python-pptx
googletrans==4.0.0-rc1
deep_translator
streamlit
